Assistant checkpoint: Add reusable content card component and update template
Assistant generated file changes:
- static/style.css: Add reusable content card styles
- templates/generation.html: Update template to use new content card structure, Update form actions to use content card footer

---

User prompt:

Are we able to fix how we create these content card containers for our forms, and potentially other pieces of data/information/menu/buttons/etc. so that we are able to simply have it be this reliable piece we can always come back to when we want to create new UI sections.

As a note as well. The content card for a lot of these forms extend up to the right-most side of the webpage, instead of wrapping around the content of which it stores/contains. Can we fix that?